pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zotero-mcp"
version = "0.1.0"
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
description = "A Model Context Protocol server for Zotero"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyzotero>=1.5.0",
    "mcp>=1.2.0",
    "python-dotenv>=1.0.0",
    "markitdown",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/zotero-mcp"
"Bug Tracker" = "https://github.com/yourusername/zotero-mcp/issues"

[project.scripts]
zotero-mcp = "zotero_mcp.cli:main"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
